+++ casadi/symbolic/profiling.cpp
@@ -61,12 +61,12 @@ double getRealTime( )
     /* Windows 2000 and later. ---------------------------------- */
     GetSystemTimeAsFileTime( &tm );
 #endif
-    t = ((ULONGLONG)tm.dwHighDateTime << 32) | (ULONGLONG)tm.dwLowDateTime;
-    return (double)t / 10000000.0;
+    t = (static_cast<ULONGLONG>(tm.dwHighDateTime) << 32) | (ULONGLONG)tm.dwLowDateTime;
+    return static_cast<double>(t) / 10000000.0;
 
 #elif (defined(__hpux) || defined(hpux)) || ((defined(__sun__) || defined(__sun) || defined(sun)) && (defined(__SVR4) || defined(__svr4__)))
     /* HP-UX, Solaris. ------------------------------------------ */
-    return (double)gethrtime( ) / 1000000000.0;
+    return static_cast<double>(gethrtime()) / 1000000000.0;
 
 #elif defined(__MACH__) && defined(__APPLE__)
     /* OSX. ----------------------------------------------------- */
@@ -75,11 +75,9 @@ double getRealTime( )
     {
         mach_timebase_info_data_t timeBase;
         (void)mach_timebase_info( &timeBase );
-        timeConvert = (double)timeBase.numer /
-            (double)timeBase.denom /
-            1000000000.0;
+        timeConvert = static_cast<double>(timeBase.numer)/static_cast<double>(timeBase.denom) / 1000000000.0;
     }
-    return (double)mach_absolute_time( ) * timeConvert;
+    return static_cast<double>(mach_absolute_time()) * timeConvert;
 
 #elif defined(_POSIX_VERSION)
     /* POSIX. --------------------------------------------------- */
@@ -105,8 +103,7 @@ double getRealTime( )
         const clockid_t id = (clockid_t)-1; /* Unknown. */
 #endif /* CLOCK_* */
         if ( id != (clockid_t)-1 && clock_gettime( id, &ts ) != -1 )
-            return (double)ts.tv_sec +
-                (double)ts.tv_nsec / 1000000000.0;
+          return ts.tv_sec + ts.tv_nsec/1000000000.0;
         /* Fall thru. */
     }
 #endif /* _POSIX_TIMERS */
@@ -114,7 +111,7 @@ double getRealTime( )
     /* AIX, BSD, Cygwin, HP-UX, Linux, OSX, POSIX, Solaris. ----- */
     struct timeval tm;
     gettimeofday( &tm, NULL );
-    return (double)tm.tv_sec + (double)tm.tv_usec / 1000000.0;
+    return tm.tv_sec + tm.tv_usec/1000000.0;
 #else
     return -1.0;        /* Failed. */
 #endif
